add dependencies module, unlocked startup_data.py that is used by entrypoint.sh
