+++ .env.sample
@@ -1,20 +1,20 @@
-# PostgreSQL
-DB_HOST=127.0.0.1
+#PostgreSQL
+DB_HOST=db
 DB_PORT=5432
 DB_USER=postgres
-DB_PASSWORD=Test1234!
-DB_NAME=cinema
-# JWT authentication
+DB_PASSWORD=postgres
+DB_NAME=online_cinema
+#JWT authentication
 SECRET_KEY_ACCESS=152fc61dbfa93362bea99c4115b379e3c4411680a4b46c90b8065e425d670cfa
 SECRET_KEY_REFRESH=7f01312fb2a30f17db1f2efc5c2152f931ed7e26c92144c57da00dff7bd0db63
-# Email Sender
-EMAIL_HOST=localhost
-EMAIL_PORT=1026
+#Email Sender
+EMAIL_HOST=mailhog
+EMAIL_PORT=1025
 EMAIL_HOST_USER=testuser
 EMAIL_HOST_PASSWORD=Test1234!
 FROM_EMAIL=no-reply@example.com
 APP_URL=http://127.0.0.1:8001
 #Redis
-REDIS_HOST=localhost
+REDIS_HOST=redis
 REDIS_PORT=6379
 REDIS_DB=0
