+++ src/config/__init__.py
@@ -1,7 +1,1 @@
-from src.config.dependencies import (
-    get_settings,
-    get_jwt_auth_manager,
-    get_email_sender,
-    get_redis_client,
-)
 from src.config.config import BaseAppSettings
