+++ src/database/session.py
@@ -1,7 +1,7 @@
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker, Session
 
-from src.config import get_settings
+from src.dependencies import get_settings
 
 settings = get_settings()
