+++ src/dependencies/__init__.py
@@ -1,0 +1,8 @@
+from src.dependencies.config import get_settings, get_email_sender, get_redis_client
+from src.dependencies.auth import (
+    get_jwt_auth_manager,
+    get_token,
+    get_current_user,
+    get_current_active_user,
+    check_admin_role,
+)
