+++ src/dependencies/auth.py
@@ -2,10 +2,22 @@ from fastapi import Request, HTTPException, status, Depends
 from redis import Redis
 from sqlalchemy.orm import Session
 
-from src.config import get_redis_client, get_jwt_auth_manager
 from src.database import UserModel
 from src.database import get_db
+from src.dependencies.config import get_settings, get_redis_client
+from src.config import BaseAppSettings
 from src.security.token_manager import JWTManager
+
+
+def get_jwt_auth_manager(
+    settings: BaseAppSettings = Depends(get_settings),
+) -> JWTManager:
+
+    return JWTManager(
+        secret_key_access=settings.SECRET_KEY_ACCESS,
+        secret_key_refresh=settings.SECRET_KEY_REFRESH,
+        algorithm=settings.ALGORITHM,
+    )
 
 
 def get_token(request: Request) -> str:
