+++ src/dependencies/config.py
@@ -1,23 +1,12 @@
 from fastapi import Depends
 from redis import Redis
 
-from src.security.token_manager import JWTManager
 from src.services import EmailSender
 from src.config.config import Settings, BaseAppSettings
 
 
 def get_settings() -> Settings:
     return Settings()
-
-
-def get_jwt_auth_manager(
-    settings: BaseAppSettings = Depends(get_settings),
-) -> JWTManager:
-    return JWTManager(
-        secret_key_access=settings.SECRET_KEY_ACCESS,
-        secret_key_refresh=settings.SECRET_KEY_REFRESH,
-        algorithm=settings.ALGORITHM,
-    )
 
 
 def get_email_sender(settings: BaseAppSettings = Depends(get_settings)) -> EmailSender:
