+++ src/routes/profiles.py
@@ -7,11 +7,11 @@ from fastapi.params import Depends
 from sqlalchemy.exc import SQLAlchemyError
 from sqlalchemy.orm import Session
 
-from src.config import get_settings
+from src.dependencies import get_settings
 from src.database import GenderEnum, UserProfileModel, UserModel
 from src.database import get_db
 from src.schemas.profiles import ProfileSchema
-from src.security.dependencies import get_current_user, get_current_active_user
+from src.dependencies import get_current_user, get_current_active_user
 from src.validation import (
     validate_name,
     validate_birth_date,
